Code improvement using factory design pattern
